Added more docstrings, edited setup.py

+++ discord/ext/forms/form.py
@@ -2,6 +2,7 @@ import discord
 from discord.ext import commands
 import re
 from typing import List
+import json
 
 class Form(object):
     """The basic form object.
@@ -36,11 +37,20 @@ class Form(object):
         self.timeout = timeout
 
     def set_tries(self,tries:int) -> None:
+        """Set the amount of tries that are allowed during input validation. Defaults to 3.
+
+        Parameters
+        ----------
+        tries : int
+            The number of tries to set.
+        """
         int(tries)
         self._tries = tries
 
     def add_question(self,question,key:str=None,qtype=None) -> List[dict]:
         """Adds a question to the form.
+        The valid qtypes are:
+        `invite`,`channel`,`user`,`member`,`role`, and `category`
 
         Parameters
         ----------
@@ -68,7 +78,7 @@ class Form(object):
         if not key:
             key = question
         valid_qtypes = ['invite','channel','user','member','role','category']
-        dictionary = {'res':None,'type':None}
+        dictionary = {'res':None,'type':None,'question':question}
         if qtype:
             if qtype.lower() not in valid_qtypes:
                 raise InvalidFormType(f"Type '{qtype}' is invalid!")
@@ -76,50 +86,52 @@ class Form(object):
             dictionary['type'] = qtype
             self._tries = 3
 
-        self._questions[key] = None
+        self._questions[key] = dictionary
+        self.set_incorrect_message('You answered incorrectly too many times. Please try again.')
+        self.set_retry_message(f'Please try again.')
         return self._questions
 
     async def validate_input(self,qtype,answer):
         if qtype.lower() == 'invite':
             try:
-                await commands.InviteConverter.convert(self._ctx,answer)
-                return True
-            except:
+                invite = await commands.InviteConverter().convert(self._ctx,answer)
+                return invite
+            except Exception as e:
                 return False
         elif qtype.lower() == 'channel':
             try:
-                await commands.TextChannelConverter.convert(self._ctx,answer)
-                return True
+                channel = await commands.TextChannelConverter().convert(self._ctx,answer)
+                return channel
             except:
                 return False
         elif qtype.lower() == 'user':
             try:
-                await commands.UserConverter.convert(self._ctx,answer)
-                return True
+                user = await commands.UserConverter().convert(self._ctx,answer)
+                return user
             except:
                 return False
         elif qtype.lower() == 'member':
             try:
-                await commands.MemberConverter.convert(self._ctx,answer)
-                return True
+                member = await commands.MemberConverter().convert(self._ctx,answer)
+                return member
             except:
                 return False
         elif qtype.lower() == 'role':
             try:
-                await commands.RoleConverter.convert(self._ctx,answer)
-                return True
+                role = await commands.RoleConverter().convert(self._ctx,answer)
+                return role
             except:
                 return False
         elif qtype.lower() == 'category':
             try:
-                await commands.CategoryChannelConverter.convert(self._ctx,answer)
-                return True
+                category = await commands.CategoryChannelConverter().convert(self._ctx,answer)
+                return category
             except:
                 return False
         elif qtype.lower() == 'emoji':
             try:
-                await commands.EmojiConverter.convert(self._ctx,answer)
-                return True
+                emoji = await commands.EmojiConverter().convert(self._ctx,answer)
+                return emoji
             except:
                 return False
         else:
@@ -157,7 +169,7 @@ class Form(object):
         message : str
             The message to be set.
         """
-        self._incorrectmsg = message
+        self._retrymsg = message
 
     def set_incorrect_message(self,message:str):
         """Sets the message to send if input validation fails and there are no more tries left..
@@ -181,7 +193,7 @@ class Form(object):
         color = await commands.ColourConverter().convert(self._ctx,color)
         self._color = color
 
-    async def start(self,channel=None) -> List[dict]:
+    async def start(self,channel=None) -> dict:
         """Starts the form in the current channel.
 
         Parameters
@@ -191,22 +203,38 @@ class Form(object):
 
         Returns
         -------
-        List[dict]
-            [description]
+        dict
+            A dictionary containing each question and its data.
+            Example:
+            ::
+
+            {'Key Specified':{
+                    'res':'answer goes here',
+                    'type':'input validation type'
+                    'question':'Question here'
+                },
+            'Next Key Specified':{
+                'res':'answer goes here',
+                'type':'input validation type'
+                'question':'Question here'
+            }
         """
         elist = []
 
         if not channel:
             channel = self._ctx.channel
 
-        for q in self._questions.keys():
-            embed=discord.Embed(description=q,color=self._color)
-            embed.set_author(name=f"{self.title}: {self._questions.index(q)+1}/{len(self._questions)}",icon_url=self._bot.user.avatar_url)
+        qlist = []
+        for n, q in enumerate(self._questions.values()):
+            embed=discord.Embed(description=q['question'],color=self._color)
+
+            embed.set_author(name=f"{self.title}: {n+1}/{len(self._questions)}",icon_url=self._bot.user.avatar_url)
 
             if self.color:
                 embed.color=self.color
 
             elist.append(embed)
+
 
         prompt = None
 
@@ -222,20 +250,44 @@ class Form(object):
 
             def check(m):
                 return m.channel == prompt.channel and m.author == self._ctx.author
+            question = None
+            for x in self._questions.keys():
+                nx = self._questions[x]
+                if nx['question'] == embed.description:
+                    question = x
 
-            question = self._questions[elist.index(embed)]
 
             msg = await self._bot.wait_for('message',check=check,timeout=self.timeout)
             ans = msg.content
-            await msg.delete()
+            if self.editanddelete:
+                await msg.delete()
             if self._tries:
-                if 'qtype' in question.keys(): # TODO: Add input validation
-                    correct = await self.validate_input(question['qtype',ans])
-                    if correct:
-                        self._questions[question] = ans
-                    else:
-                        self._tries -= 1
-
+                print("Tries found...")
+                key = question
+                question = self._questions[question]
+                if 'type' in question.keys():
+                    while True:
+                        result = await self.validate_input(question['type'],ans)
+                        if result:
+                            print("Input validation worked!")
+                            nx = question
+                            nx['res'] = result
+                            self._questions[key] = nx
+                            print(self._questions)
+                                    #self._questions[x] = ans
+                            #self._questions[question] = ans
+                            break
+                        else:
+                            await channel.send(self._retrymsg+f" You have `{self._tries}` remaining.")
+                            print("Input validation failed...")
+                            msg = await self._bot.wait_for('message',check=check,timeout=self.timeout)
+                            ans = msg.content
+                            if self.editanddelete:
+                                await msg.delete()
+            else:
+                self._questions[question] = ans
+        for i in self._questions.keys():
+            self._questions[i] = self._questions[i]
         return self._questions
 
 class InvalidColor(Exception):
